bug fix: delete article

+++ src/FitoGraph.Api/Areas/Admin/Handlers/DeleteArticleCommandHandler.cs
@@ -34,7 +34,8 @@ namespace FitoGraph.Api.Areas.Admin.Handlers
                     result.Message = "Article doesn't exists";
                     return result;
                 }
-                _dbContext.TArticleSport.RemoveRange(tData.TArticleSports);
+                if (tData.TArticleSports != null && tData.TArticleSports.Any())
+                    _dbContext.TArticleSport.RemoveRange(tData.TArticleSports);
                 _dbContext.TArticle.Remove(tData);
                 await _dbContext.SaveChangesAsync();
                 result.Status = true;
